server/requirements.txt
fastapi==0.68.1
fastapi-sqlalchemy==0.2.1
pydantic==1.8.2
alembic==1.7.3
psycopg2==2.9.1
uvicorn==0.15.0
importlib-resources==5.2.2
python-dotenv==0.19.0
bcrypt==3.2.0
pytest==6.2.5
requests==2.26.0
ipdb==0.13.9